Update BiasDetector.py
Messing around with some variables, added more stop words as there is very likely a huge subset of junk that can be present intisde the wide variety of news documents we are parsing. Also increased the number of topics, since it seemed like, with only 20 topics, our topic vectors were essentailly 99% one topic and equally dsitributed probability of the rest, which is p oor for classification.